// src/api/generated/models/pageResultCommentVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CommentVO } from './commentVO';

export interface PageResultCommentVO {
  list?: CommentVO[];
  total?: string;
  pageNum?: number;
  pageSize?: number;
  totalPages?: number;
  hasNext?: boolean;
  hasPrevious?: boolean;
}